+++ kit/dapp/src/app/[locale]/(private)/assets/[assettype]/[address]/_components/manage-dropdown/grant-role-form/form.tsx
@@ -2,15 +2,15 @@
 
 import { Form } from '@/components/blocks/form/form';
 import { FormSheet } from '@/components/blocks/form/form-sheet';
-import { bondGrantRoleAction, cryptoCurrencyGrantRoleAction, equityGrantRoleAction, fundGrantRoleAction, stablecoinGrantRoleAction, tokenizedDepositGrantRoleAction, type GrantRoleActionType } from '@/lib/mutations/asset/access-control/grant-role/grant-role-action';
+import { grantRole } from '@/lib/mutations/asset/access-control/grant-role/grant-role-action';
 import { GrantRoleSchema } from '@/lib/mutations/asset/access-control/grant-role/grant-role-schema';
+import type { AssetType } from '@/lib/utils/zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useTranslations } from 'next-intl';
 import type { Address } from 'viem';
 import { AdminAddress } from './steps/address';
 import { AdminRoles } from './steps/roles';
 import { Summary } from './steps/summary';
-import type { AssetType } from '../../../../types';
 
 interface GrantRoleFormProps {
   address: Address;
@@ -26,24 +26,7 @@ export function GrantRoleForm({
   assettype,
 }: GrantRoleFormProps) {
   const t = useTranslations('private.assets.details.forms.grant-role');
-  const getGrantRoleAction = (assettype: AssetType) => {
-    switch (assettype) {
-      case 'bonds':
-        return bondGrantRoleAction;
-      case 'equities':
-        return equityGrantRoleAction;
-      case 'funds':
-        return fundGrantRoleAction;
-      case 'tokenizeddeposits':
-        return tokenizedDepositGrantRoleAction;
-      case 'cryptocurrencies':
-        return cryptoCurrencyGrantRoleAction;
-      case 'stablecoins':
-        return stablecoinGrantRoleAction;
-      default:
-        throw new Error(`Invalid asset type: ${assettype}`);
-    }
-  };
+
   return (
     <FormSheet
       open={open}
@@ -52,7 +35,7 @@ export function GrantRoleForm({
       description={t('description')}
     >
       <Form
-        action={getGrantRoleAction(assettype)}
+        action={grantRole}
         resolver={zodResolver(GrantRoleSchema)}
         onOpenChange={onOpenChange}
         buttonLabels={{
@@ -60,6 +43,7 @@ export function GrantRoleForm({
         }}
         defaultValues={{
           address,
+          assettype,
           roles: {
             DEFAULT_ADMIN_ROLE: false,
             SUPPLY_MANAGEMENT_ROLE: false,
