+++ kit/dapp/src/app/[locale]/(private)/assets/[assettype]/[address]/_components/manage-dropdown/manage-dropdown.tsx
@@ -7,7 +7,6 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from '@/components/ui/dropdown-menu';
-import { bondGrantRoleAction } from '@/lib/mutations/asset/access-control/grant-role/grant-role-action';
 import type { getAssetDetail } from '@/lib/queries/asset-detail';
 import type { getBondDetail } from '@/lib/queries/bond/bond-detail';
 import type { AssetType } from '@/lib/utils/zod';
