+++ kit/dapp/src/components/blocks/asset-revoke-all-permissions/form.tsx
@@ -3,13 +3,14 @@
 import { Form } from '@/components/blocks/form/form';
 import { FormSheet } from '@/components/blocks/form/form-sheet';
 import type { Role } from '@/lib/config/roles';
-import { bondRevokeRoleAction, cryptoCurrencyRevokeRoleAction, equityRevokeRoleAction, fundRevokeRoleAction, stableCoinRevokeRoleAction, tokenizedDepositRevokeRoleAction, type RevokeRoleActionType } from '@/lib/mutations/asset/access-control/revoke-role/revoke-role-action';
+import { revokeRole } from '@/lib/mutations/asset/access-control/revoke-role/revoke-role';
+
 import { RevokeRoleSchema } from '@/lib/mutations/asset/access-control/revoke-role/revoke-role-schema';
+import type { AssetType } from '@/lib/utils/zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useTranslations } from 'next-intl';
 import type { Address } from 'viem';
 import { Summary } from './steps/summary';
-import type { AssetType } from '@/app/[locale]/(private)/assets/[assettype]/types';
 
 export interface RevokeAllPermissionsFormProps {
   address: Address;
@@ -36,24 +37,6 @@ export function RevokeAllPermissionsForm({
     'private.assets.details.permissions.revoke-all-form'
   );
 
-  const getRevokeRoleAction = (assettype: AssetType) => {
-    switch (assettype) {
-      case 'bonds':
-        return bondRevokeRoleAction;
-      case 'equities':
-        return equityRevokeRoleAction;
-      case 'funds':
-        return fundRevokeRoleAction;
-      case 'tokenizeddeposits':
-        return tokenizedDepositRevokeRoleAction;
-      case 'cryptocurrencies':
-        return cryptoCurrencyRevokeRoleAction;
-      case 'stablecoins':
-        return stableCoinRevokeRoleAction;
-      default:
-        throw new Error(`Invalid asset type: ${assettype}`);
-    }
-  };
   return (
     <FormSheet
       open={open}
@@ -63,7 +46,7 @@ export function RevokeAllPermissionsForm({
       description={t('description')}
     >
       <Form
-        action={getRevokeRoleAction(assettype)}
+        action={revokeRole}
         resolver={zodResolver(RevokeRoleSchema)}
         onOpenChange={onOpenChange}
         buttonLabels={{
@@ -72,6 +55,7 @@ export function RevokeAllPermissionsForm({
         defaultValues={{
           address,
           userAddress: account,
+          assettype,
           roles: currentRoles.reduce(
             (acc, role) => {
               acc[role] = true;
