+++ kit/dapp/src/lib/mutations/asset/access-control/revoke-role/revoke-role-schema.ts
@@ -13,6 +13,7 @@ export const RevokeRoleSchema = z.object({
   roles: z.roles(),
   userAddress: z.address(),
   pincode: z.pincode(),
+  assettype: z.assetType(),
 });
 
 export type RevokeRoleInput = ZodInfer<typeof RevokeRoleSchema>;
