+++ kit/dapp/src/lib/mutations/asset/access-control/update-role/update-role-schema.ts
@@ -13,6 +13,7 @@ export const UpdateRolesSchema = z.object({
   roles: z.roles(),
   userAddress: z.address(),
   pincode: z.pincode(),
+  assettype: z.assetType(),
 });
 
 export type UpdateRolesInput = ZodInfer<typeof UpdateRolesSchema>;
