--- kit/dapp/src/lib/mutations/asset/access-control/update-role/update-roles-action.ts
@@ -1,59 +1,0 @@
-/* eslint-disable @typescript-eslint/no-duplicate-type-constituents */
-
-'use server';
-
-import {
-  BondGrantRole,
-  CryptoCurrencyGrantRole,
-  EquityGrantRole,
-  FundGrantRole,
-  StableCoinGrantRole,
-  TokenizedDepositGrantRole,
-} from '../grant-role/grant-role';
-import {
-  BondRevokeRole,
-  CryptoCurrencyRevokeRole,
-  EquityRevokeRole,
-  FundRevokeRole,
-  StableCoinRevokeRole,
-  TokenizedDepositRevokeRole,
-} from '../revoke-role/revoke-role';
-import { getUpdateRolesAction } from './update-roles';
-
-export const stableCoinUpdatePermissionsAction = getUpdateRolesAction({
-  grantRoleMutation: StableCoinGrantRole,
-  revokeRoleMutation: StableCoinRevokeRole,
-});
-
-export const bondUpdatePermissionsAction = getUpdateRolesAction({
-  grantRoleMutation: BondGrantRole,
-  revokeRoleMutation: BondRevokeRole,
-});
-
-export const cryptoCurrencyUpdatePermissionsAction = getUpdateRolesAction({
-  grantRoleMutation: CryptoCurrencyGrantRole,
-  revokeRoleMutation: CryptoCurrencyRevokeRole,
-});
-
-export const fundUpdatePermissionsAction = getUpdateRolesAction({
-  grantRoleMutation: FundGrantRole,
-  revokeRoleMutation: FundRevokeRole,
-});
-
-export const equityUpdatePermissionsAction = getUpdateRolesAction({
-  grantRoleMutation: EquityGrantRole,
-  revokeRoleMutation: EquityRevokeRole,
-});
-
-export const tokenizedDepositUpdatePermissionsAction = getUpdateRolesAction({
-  grantRoleMutation: TokenizedDepositGrantRole,
-  revokeRoleMutation: TokenizedDepositRevokeRole,
-});
-
-export type UpdateRolesActionType =
-  | typeof stableCoinUpdatePermissionsAction
-  | typeof bondUpdatePermissionsAction
-  | typeof cryptoCurrencyUpdatePermissionsAction
-  | typeof fundUpdatePermissionsAction
-  | typeof equityUpdatePermissionsAction
-  | typeof tokenizedDepositUpdatePermissionsAction;
